refactor: 알림 컴포넌트 정리 및 파일 구조 개선(#343)
