+++ src/components/header/components/notification-section/notificationIconClass.ts
